requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
bcrypt
pyjwt
passlib
